heroku un

+++ src/main/java/com/example/springbootmvc/controller/CompanyController.java
@@ -1,12 +1,16 @@
 package com.example.springbootmvc.controller;
 
 import com.example.springbootmvc.entity.Company;
+import com.example.springbootmvc.entity.Department;
 import com.example.springbootmvc.repository.CompanyRepository;
+import com.example.springbootmvc.repository.DepartmentRepository;
 import com.example.springbootmvc.service.CompanyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @Controller
 @RequestMapping("/company")
@@ -16,6 +20,8 @@ public class CompanyController {
     CompanyService companyService;
     @Autowired
     CompanyRepository companyRepository;
+    @Autowired
+    DepartmentRepository departmentRepository;
 
     @GetMapping
     public String getCompanyPage(Model model) {
@@ -36,7 +42,10 @@ public class CompanyController {
 
     @GetMapping("/delete/{id}") //1 45 24 90
     public String delete(@PathVariable Integer id) {
-        companyRepository.deleteById(id);
+        Optional<Department> departmentByCompany_id = departmentRepository.findDepartmentByCompany_Id(id);
+        if (departmentByCompany_id.isEmpty()) {
+            companyRepository.deleteById(id);
+        }
         return "redirect:/company";
     }
